# atlas.com/wrg/swagger.yaml
basePath: /ms/wrg/worlds
consumes:
- application/json
definitions:
  ChannelServerAttributes:
    properties:
      capacity:
        format: int64
        type: integer
        x-go-name: Capacity
      channelId:
        format: uint8
        type: integer
        x-go-name: ChannelId
      ipAddress:
        type: string
        x-go-name: IpAddress
      port:
        format: int64
        type: integer
        x-go-name: Port
      worldId:
        format: uint8
        type: integer
        x-go-name: WorldId
    type: object
    x-go-package: _/Users/jtumidanski/Documents/Projects/atlas/atlas-wrg2/atlas.com/wrg/attributes
  ChannelServerData:
    properties:
      attributes:
        $ref: '#/definitions/ChannelServerAttributes'
      id:
        type: string
        x-go-name: Id
      type:
        type: string
        x-go-name: Type
    type: object
    x-go-package: _/Users/jtumidanski/Documents/Projects/atlas/atlas-wrg2/atlas.com/wrg/attributes
  worldServerAttributes:
    properties:
      capacityStatus:
        format: int64
        type: integer
        x-go-name: CapacityStatus
      eventMessage:
        type: string
        x-go-name: EventMessage
      flag:
        format: int64
        type: integer
        x-go-name: Flag
      message:
        type: string
        x-go-name: Message
      name:
        type: string
        x-go-name: Name
      recommended:
        type: boolean
        x-go-name: Recommended
      recommendedMessage:
        type: string
        x-go-name: RecommendedMessage
    type: object
    x-go-name: WorldAttributes
    x-go-package: _/Users/jtumidanski/Documents/Projects/atlas/atlas-wrg2/atlas.com/wrg/attributes
  worldServerData:
    properties:
      attributes:
        $ref: '#/definitions/worldServerAttributes'
      id:
        type: string
        x-go-name: Id
      type:
        type: string
        x-go-name: Type
    type: object
    x-go-name: WorldData
    x-go-package: _/Users/jtumidanski/Documents/Projects/atlas/atlas-wrg2/atlas.com/wrg/attributes
info:
  description: Documentation for World API
  title: of World API
  version: 1.0.0
paths:
  /channelServers:
    get:
      description: Return a list of channel servers in the world registry
      operationId: getChannelServers
      responses:
        "200":
          $ref: '#/responses/channelServersResponse'
      tags:
      - channelServers
    post:
      description: Registers a channel server to the world registry
      operationId: register
      responses:
        "200":
          $ref: '#/responses/channelServerResponse'
      tags:
      - channelServers
  /channelServers/{id}:
    delete:
      description: Removes channel server registration from the world registry
      operationId: unregister
      parameters:
      - description: The id of the channel for which the operation relates
        format: int64
        in: path
        minimum: 1
        name: channelId
        required: true
        type: integer
        x-go-name: ChannelId
      responses:
        "204":
          $ref: '#/responses/noContentResponse'
      tags:
      - channelServers
  /worlds:
    get:
      operationId: getWorlds
      responses:
        "200":
          $ref: '#/responses/worldServersResponse'
      summary: Retrieves all world server information.
      tags:
      - worlds
  /worlds/{worldId}:
    get:
      operationId: getWorld
      parameters:
      - description: The id of the world for which the operation relates
        format: int64
        in: path
        minimum: 0
        name: worldId
        required: true
        type: integer
        x-go-name: WorldId
      responses:
        "200":
          $ref: '#/responses/worldServerResponse'
        "404":
          $ref: '#/responses/notFoundResponse'
      summary: Retrieves world server information.
      tags:
      - worlds
  /worlds/{worldId}/channels/{channelId}:
    get:
      operationId: getChannelServer
      parameters:
      - description: The id of the world for which the operation relates
        format: int64
        in: path
        minimum: 0
        name: worldId
        required: true
        type: integer
        x-go-name: WorldId
      - description: The id of the channel for which the operation relates
        format: int64
        in: path
        minimum: 1
        name: channelId
        required: true
        type: integer
        x-go-name: ChannelId
      responses:
        "200":
          $ref: '#/responses/channelServerResponse'
        "404":
          $ref: '#/responses/notFoundResponse'
      summary: Retrieves channel server information for a worlds channel.
      tags:
      - worlds
produces:
- application/json
responses:
  channelServerResponse:
    description: A channelServer
    schema:
      $ref: '#/definitions/ChannelServerData'
  channelServersResponse:
    description: A list of channelServers
    schema:
      items:
        $ref: '#/definitions/ChannelServerData'
      type: array
  noContentResponse:
    description: No content is returned by this API endpoint
  notFoundResponse:
    description: The requested resource was not found
  worldServerResponse:
    description: A worldServer
    headers:
      data:
        description: A world server
    schema:
      $ref: '#/definitions/worldServerData'
  worldServersResponse:
    description: A list of worldServers
    headers:
      data:
        description: All world servers
        items:
          $ref: '#/definitions/worldServerData'
        type: array
schemes:
- http
swagger: "2.0"